# backend/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agency Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Campaign updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /campaigns/{id}/pause:
    post:
      operationId: pauseCampaign
      tags: [campaigns]
      summary: Pause a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/resume:
    post:
      operationId: resumeCampaign
      tags: [campaigns]
      summary: Resume a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  # Adsets
...
  /ai-insights:
    get:
      operationId: listAiInsights
      tags: [ai-insights]
      summary: List AI insights
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of AI insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiInsight"
    post:
      operationId: generateAiInsight
      tags: [ai-insights]
      summary: Generate an AI insight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiInsightRequest"
      responses:
        "201":
          description: AI insight generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiInsight"
  /ai-insights/{id}:
    get:
      operationId: getAiInsight
      tags: [ai-insights]
      summary: Get an AI insight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI Insight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiInsight"
    delete:
      operationId: deleteAiInsight
      tags: [ai-insights]
      summary: Delete an AI insight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Notifications
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    Campaign:
      type: object
      required: [id, name, clientId, status, objective, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        metaCampaignId:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, paused, completed, draft]
        objective:
          type: string
        dailyBudget:
          type: ["number", "null"]
        totalBudget:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        spend:
          type: ["number", "null"]
        impressions:
          type: ["integer", "null"]
        clicks:
          type: ["integer", "null"]
        conversions:
          type: ["integer", "null"]
        roas:
          type: ["number", "null"]
        createdAt:
          type: string
    CampaignInput:
      type: object
      required: [name, clientId, objective]
      properties:
        name:
          type: string
        clientId:
          type: integer
        objective:
          type: string
        dailyBudget:
          type: number
        totalBudget:
          type: number
        startDate:
          type: string
        endDate:
          type: string
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, paused, completed, draft]
        objective:
          type: string
        dailyBudget:
          type: number
        totalBudget:
          type: number
        startDate:
          type: string
        endDate:
          type: string

    # Adset
...
    AiInsight:
      type: object
      required: [id, type, content, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        campaignId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [campaign_summary, performance_analysis, weekly_insights, loss_debrief, call_coaching, revenue_analysis, recommendations]
        content:
          type: string
        prompt:
          type: ["string", "null"]
        model:
          type: string
        createdAt:
          type: string
    AiInsightRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [campaign_summary, performance_analysis, weekly_insights, loss_debrief, call_coaching, revenue_analysis, recommendations]
        clientId:
          type: integer
        campaignId:
          type: integer
        context:
          type: string

    # Notification
...